requirements.txt
cookiecutter
coverage
detect-secrets==1.0.3
govuk-tech-docs-sphinx-theme
myst-parser
pre-commit
pytest
pytest-cookies
pytest-mock
pytest-xdist
python-dotenv
Sphinx
toml
